First compilation with Qt5.12.0 for Windows

+++ src/sql/qxtsqlthreadmanager.cpp
@@ -100,7 +100,7 @@ QxtSqlThreadManager::QxtSqlThreadManager(const QString &masterName)
 {
     // Build a name for the new connection
     this->name = QString("$qxt$tc_%1$%2")
-	.arg((unsigned long)(void*)QThread::currentThreadId(), 0, 36)
+	.arg((intptr_t)(void*)QThread::currentThreadId(), 0, 36)
 	.arg(masterName);
     // Clone the primary thread's connection
     Q_ASSERT(QSqlDatabase::contains(masterName));
