+++ tools/3rdparty/qdoc3/codemarker.cpp
@@ -176,7 +176,7 @@ const Node *CodeMarker::nodeForString(const QString& string)
 QString CodeMarker::stringForNode(const Node *node)
 {
     if (sizeof(const Node *) == sizeof(ulong)) {
-        return QString::number(reinterpret_cast<ulong>(node));
+        return QString::number(reinterpret_cast<intptr_t>(node));
     } else {
         return QString::number(reinterpret_cast<qulonglong>(node));
     }
